+++ lib/src/network/cnn.dart
@@ -1,1 +1,2 @@
+/// Class that represent the convolutional neural network (CNN)
 class CNN {}